Move Target, Range & Category enums to their own files
