+++ PMDMMO-common/src/com/darkxell/common/dungeon/DungeonExploration.java
@@ -26,8 +26,8 @@ import com.darkxell.common.mission.Mission;
 import com.darkxell.common.player.Player;
 import com.darkxell.common.pokemon.DungeonPokemon;
 import com.darkxell.common.pokemon.DungeonPokemon.DungeonPokemonType;
-import com.darkxell.common.registry.Registries;
 import com.darkxell.common.pokemon.Pokemon;
+import com.darkxell.common.registry.Registries;
 import com.darkxell.common.status.AppliedStatusCondition;
 import com.darkxell.common.util.Direction;
 import com.darkxell.common.util.Logger;
