+++ PMDMMO-common/src/com/darkxell/common/event/item/ItemSelectionEvent.java
@@ -7,8 +7,8 @@ import com.darkxell.common.event.Event;
 import com.darkxell.common.event.EventSource;
 import com.darkxell.common.item.Item;
 import com.darkxell.common.item.ItemContainer;
-import com.darkxell.common.item.ItemStack;
 import com.darkxell.common.item.ItemContainer.ItemContainerType;
+import com.darkxell.common.item.ItemStack;
 import com.darkxell.common.pokemon.DungeonPokemon;
 import com.darkxell.common.pokemon.Pokemon;
 import com.darkxell.common.util.Communicable;
