+++ PMDMMO-common/src/com/darkxell/common/move/MoveCategory.java
@@ -1,0 +1,13 @@
+package com.darkxell.common.move;
+
+import com.darkxell.common.util.language.Message;
+
+public enum MoveCategory {
+    Physical,
+    Special,
+    Status;
+
+    public Message getName() {
+        return new Message("move.info.category." + this.name());
+    }
+}
