+++ PMDMMO-common/src/com/darkxell/common/move/behavior/MoveTargetSelector.java
@@ -7,7 +7,7 @@ import com.darkxell.common.dungeon.floor.Tile;
 import com.darkxell.common.dungeon.floor.TileType;
 import com.darkxell.common.dungeon.floor.room.Room;
 import com.darkxell.common.move.Move;
-import com.darkxell.common.move.Move.MoveRange;
+import com.darkxell.common.move.MoveRange;
 import com.darkxell.common.pokemon.DungeonPokemon;
 import com.darkxell.common.pokemon.DungeonPokemon.DungeonPokemonType;
 import com.darkxell.common.status.StatusConditions;
