+++ PMDMMO-common/src/com/darkxell/common/pokemon/ability/AbilityOnHit.java
@@ -7,7 +7,7 @@ import com.darkxell.common.event.Event;
 import com.darkxell.common.event.move.MoveSelectionEvent.MoveUse;
 import com.darkxell.common.event.pokemon.DamageDealtEvent;
 import com.darkxell.common.event.pokemon.TriggeredAbilityEvent;
-import com.darkxell.common.move.Move.MoveCategory;
+import com.darkxell.common.move.MoveCategory;
 import com.darkxell.common.pokemon.DungeonPokemon;
 
 public abstract class AbilityOnHit extends Ability {
