+++ PMDMMO-common/src/com/darkxell/common/status/conditions/SetAccuracyStatusCondition.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 
 import com.darkxell.common.dungeon.floor.Floor;
 import com.darkxell.common.event.Event;
-import com.darkxell.common.event.move.MoveUseEvent;
 import com.darkxell.common.event.move.MoveSelectionEvent.MoveUse;
-import com.darkxell.common.pokemon.DungeonPokemon;
+import com.darkxell.common.event.move.MoveUseEvent;
 import com.darkxell.common.pokemon.BaseStats.Stat;
+import com.darkxell.common.pokemon.DungeonPokemon;
 import com.darkxell.common.status.StatusCondition;
 
 public class SetAccuracyStatusCondition extends StatusCondition {
