+++ PMDMMO-common/test/com/darkxell/common/event/item/ItemEventTest.java
@@ -13,8 +13,8 @@ import com.darkxell.common.event.Event;
 import com.darkxell.common.event.pokemon.DamageDealtEvent;
 import com.darkxell.common.item.Item;
 import com.darkxell.common.item.Item.ItemAction;
-import com.darkxell.common.registry.Registries;
 import com.darkxell.common.item.ItemStack;
+import com.darkxell.common.registry.Registries;
 import com.darkxell.common.testutils.AssertUtils;
 import com.darkxell.common.util.Direction;
