+++ PMDMMO-common/test/com/darkxell/common/move/MoveRangeTest.java
@@ -18,8 +18,6 @@ import com.darkxell.common.dungeon.data.FloorSet;
 import com.darkxell.common.dungeon.floor.layout.Layout;
 import com.darkxell.common.event.Event;
 import com.darkxell.common.event.move.MoveSelectionEvent;
-import com.darkxell.common.move.Move.MoveRange;
-import com.darkxell.common.move.Move.MoveTarget;
 import com.darkxell.common.pokemon.DungeonPokemon;
 import com.darkxell.common.pokemon.LearnedMove;
 import com.darkxell.common.pokemon.Pokemon;
