+++ PMDMMO-common/test/com/darkxell/common/move/effects/MoveEffectsTest.java
@@ -31,9 +31,9 @@ import com.darkxell.common.event.stats.PPChangedEvent;
 import com.darkxell.common.event.stats.StatChangedEvent;
 import com.darkxell.common.item.ItemStack;
 import com.darkxell.common.move.Move;
-import com.darkxell.common.move.Move.MoveCategory;
-import com.darkxell.common.move.Move.MoveRange;
 import com.darkxell.common.move.MoveBuilder;
+import com.darkxell.common.move.MoveCategory;
+import com.darkxell.common.move.MoveRange;
 import com.darkxell.common.pokemon.BaseStats.Stat;
 import com.darkxell.common.pokemon.LearnedMove;
 import com.darkxell.common.pokemon.PokemonType;
